Refactor: changed import in some Dtos, since it broke when moving with git mv.
