+++ src/main/java/pl/lodz/p/it/eduvirt/controller/ResourceGroupVmController.java
@@ -4,7 +4,7 @@ import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import pl.lodz.p.it.eduvirt.dto.VmDto;
+import pl.lodz.p.it.eduvirt.dto.vm.VmDto;
 import pl.lodz.p.it.eduvirt.dto.resource_group.AddVmDto;
 import pl.lodz.p.it.eduvirt.entity.eduvirt.ResourceGroup;
 import pl.lodz.p.it.eduvirt.mappers.VmMapper;
