+++ src/main/java/pl/lodz/p/it/eduvirt/dto/vm/VmDto.java
@@ -1,4 +1,4 @@
-package pl.lodz.p.it.eduvirt.dto;
+package pl.lodz.p.it.eduvirt.dto.vm;
 
 public record VmDto(String id, String name) {
 }
